small architecture refactoring
